refactor(gateway): extract cached agent state mixin
